Rename norm -> normalization
