+++ external/fv3fit/fv3fit/emulation/layers/__init__.py
@@ -1,4 +1,4 @@
-from .norm import (
+from .normalization import (
     StandardNormLayer,
     StandardDenormLayer,
     MaxFeatureStdNormLayer,
